requirements.txt
pillow==10.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
uuid==1.30
